Split the interactive endpoints between build and run
