+++ common/controllers/index.js
@@ -1,6 +1,6 @@
 const { name, version, description } = require('../../package.json');
-const { sampleView: releaseTypeSelection } = require('../services/slack/surfaces/modals/publish-release/release-type-selection');
-const { sampleView: releaseTagSelection } = require('../services/slack/surfaces/modals/deploy-release/release-tag-selection');
+const { sampleView: releaseTypeSelection } = require('../../build/services/slack/surfaces/modals/publish-release/release-type-selection');
+const { sampleView: releaseTagSelection } = require('../../run/services/slack/surfaces/modals/deploy-release/release-tag-selection');
 const { sampleView: releaseDeploymentConfirmation } = require('../../run/services/slack/surfaces/modals/deploy-release/release-deployment-confirmation');
 const { sampleView: releasePublicationConfirmation } = require('../../build/services/slack/surfaces/modals/publish-release/release-publication-confirmation');
