+++ run/routes/slack.js
@@ -1,0 +1,11 @@
+const { slackConfig } = require('../../common/config');
+const slackbotController = require('../controllers/slack');
+
+module.exports = [
+  {
+    method: 'POST',
+    path: '/run/slack/interactive-endpoint',
+    handler: slackbotController.interactiveEndpoint,
+    config: slackConfig
+  },
+];
